Accept .html at the end of post URLs
This is much more complicated than it sounds at first. Mike wants the
.html and I don't. Neither of us sees a valid reason for serving posts
in anything but the default flavour, but it sounds like something that
we should support. So there are three sets of people to make happy, not
counting the people reading the code.

So: we accept @post_flavours as extensions at the end of post URLs, and
Mike is happy. I can set @flavours and @post_flavours to be empty, but
set @index_flavours to (html,rss), and I'm happy. Other people can set
either @flavours or @{post,index}_flavours as they wish, and they are
happy enough as far as I'm concerned.

And the code doesn't look too bad.