Accept tagless postings.
When several postings were git-added at once, post-commit would
use the previous posting's tags as the default value for each
posting (except the first). This commit changes the default to be
an empty set of tags.

The meta data line is left alone, since the script earlier ensures that
each posting contains a date.