Update template to HTML5 and provide better CSS hooks.
Doctype is now HTML5.

A new meta persuades mobile phones to trust the CSS. Without this, a
phones would pretend to have a 1024-pixel screen and scale the page down
as needed. With this, the phone requires the CSS to adapt to the screen's
size.

A new CSS file is used, loathsxome.css.

The generated document uses the new HTML5 tags header, article and footer.
The way Loathsxome generates dates doesn't quite fit into that scheme, so
they're given a <div class>.

The old use of h1, h2 and h3 disappears. Now there's only a h1 for each
posting, and the CSS has to use better selectors. What used to be 'h2'
is now '.date', what used to be 'h3' is now 'article h1'.

The inline style for the footer is replaced by the CSS selector 'footer'.