help frontend caches
Posts do not change often, and can be cached for a long time, by
a frontend RAM cache, by a web accelerator, or by the browser. The
blog's home page may change more often, but still should be cached
by our frontend so that in case a blog is slashdotted loathsxome.cgi
will not be the bottleneck.